Add pytest configuration as util for reusability
